+++ app/src/main/java/com/maishapay/model/client/api/MaishapayAPI.java
@@ -183,7 +183,7 @@ public interface MaishapayAPI {
 
     @FormUrlEncoded
     @POST(BuildConfig.END_POINT)
-    Observable<Double> taux_du_jour(@Field("ent") String ent,
+    Observable<Float> taux_du_jour(@Field("ent") String ent,
                                      @Field("mn") String telephone);
 
     @FormUrlEncoded
